System: Boot Environments - Addtional cleanups and restructures
Final bits and pieces to bring bectl into OPNsense.
* Move BootEnvironments into Core package as this is likely a more logical place (and api endpoints look nicer, /api/core/boot_environment/)
* Rename the menu item to Snapshots, when possible we prefer single clause names and snapshots might need less explanation for non-fbsd users (eventually, we are creating snapshots here)
* Change zfs check message to a "alert alert-warning" dialog and change the text a bit
* Remove used includes [use] in php files
* Merge Menu and ACL with Core model, same as firmware

ref  https://github.com/opnsense/core/pull/7650
