--- src/opnsense/mvc/app/models/OPNsense/BootEnvironments/Menu/Menu.xml
@@ -1,5 +1,0 @@
-<menu>
-    <System>
-        <BootEnvironments VisibleName="Boot Environments" cssClass="fa fa-database fa-fw" url="/ui/bootenvironments/general"/>
-    </System>
-</menu>
