+++ src/opnsense/service/conf/actions.d/actions_bootenvironment.conf
@@ -30,7 +30,7 @@ message:Deleting boot environment
 
 [rename]
 command:/usr/local/opnsense/scripts/bootenvironments/bectl.py
-parameters: rename --beName %s --from-source %s
+parameters: rename --from-source %s --beName %s
 type:script_output
 message:Renaming boot environment
